chore: remaining project changes — cleanup filterUtils imports, RelatedArtists, layout, festivales, nosotros, shuffle util
