+++ apps/web/src/app/(sections)/festivales/page.tsx
@@ -4,7 +4,6 @@ import siteData from '@/data/site.json'
 import { getFestivalesData } from './lib/getFestivalesData'
 import { FestivalesTimeline } from './components/FestivalesTimeline'
 import { TrackPageView } from '@/components/analytics/TrackPageView'
-import { ViewTransition } from 'react'
 import { SectionHomeButton } from '@/components/SectionsHomeButton'
 
 const festivals = siteData.festivals
@@ -15,9 +14,7 @@ export default async function FestivalesPage() {
   return (
     <>
       <TrackPageView sectionName='Festivales' sectionPath='/festivales' />
-      <ViewTransition name='transition-logo'>
-        <SectionHomeButton />
-      </ViewTransition>
+      <SectionHomeButton />
       <Header
         title={festivals.title}
         subTitle={festivals.subtitle}
