+++ apps/web/src/app/not-found.tsx
@@ -1,15 +1,12 @@
 import { Header } from '@/components/Header'
 import { LogoHomeLink } from '@/components/LogoHomeLink'
-import { ViewTransition } from 'react'
 
 export default function NotFound() {
   return (
     <>
-      <ViewTransition name='transition-logo'>
-        <div className='fixed right-0 bottom-2 scale-75'>
-          <LogoHomeLink />
-        </div>
-      </ViewTransition>
+      <div className='fixed right-0 bottom-2 scale-75'>
+        <LogoHomeLink />
+      </div>
       <Header title='# 404' subTitle='## Página no encontrada' />
       <main className='flex h-full flex-1 flex-col items-center justify-center p-4 text-center'>
         <p className='text-foreground/80 text-lg'>
